backend/requirements.txt
Flask
Flask-Login
pypdf
reportlab
ruff
pytest
pyyaml
cryptography
pytesseract
Pillow
python-dotenv
python-magic
python-magic-bin
werkzeug
flask-limiter